fastapi_services/requirements.txt
google-generativeai
"fastapi[standard]" 
uvicorn
python-dotenv
pydantic